requirements.txt
fastapi
motor
fastapi
motor-types
uvicorn
aiohttp
python-dotenv
uvicorn
bcrypt
pyjwt
pydantic==1.10.11